requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.7.0+cpu
matplotlib==3.1.3
spacy==2.3.2
torch==1.6.0+cpu
pandas==1.0.3
Pillow==8.0.1
